+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 .
 
+## [8.2.0] - 2021-11-15
+
+### Added
+
+-   Sign in with Discord, Google workspaces.
+-   If `getProfileInfo` throws an error, then it is sent as a FIELD_ERROR to the frontend
+
+### Changed
+
+-   Changes `sendJSONResponse to not allow setting of result if it has already been set
+-   AWS lambda middleware will return `404` response if the API route is not served by the middleware and user has not passed a handler.
+
+## [8.1.2] - 2021-11-15
+
+-   Uses supertokens-js-override from npm
+
 ## [8.1.1] - 2021-11-08
 
 ### Changes